bajaj-insurance-bot/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pdfplumber
sentence-transformers
faiss-cpu
numpy
pymongo
python-dotenv
